requirement.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dlib @ file:///C:/Users/User/Downloads/dlib-python%20whl%20packages/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models @ file:///E:/face-attendance/my_attendance
Flask==3.0.3
flask-restx==1.3.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==74.1.2
Werkzeug==3.0.4
